requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.38.0
Pillow>=9.0.0
requests>=2.28.0
